Game finally works again... mostly
light cube is missing, but still. Now all the game data is in json files.
And we have transform types! That's great, will allow for further generalizations.
Transforms tend to be global and either apply to a subset of objects or all objects.
Global transforms are clearly their own type, and each non global transform is its own
type of transform